requirements.txt
praw==7.7.1
openai>=2.7.1
google-generativeai>=0.3.0
gspread==5.12.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
slack-sdk==3.27.0
pandas==2.2.0
requests==2.31.0
python-arango>=7.6.0
fastapi>=0.115.0
uvicorn>=0.30.0

